requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dialogflow==1.1.1
Flask==3.0.1
google-api-core==1.34.0
google-auth==2.27.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.48.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
urllib3==2.1.0
Werkzeug==3.0.1
